third_party/WebKit/LayoutTests/media/gc-while-delaying-the-load-event.html

<!DOCTYPE html>
<title>GC while the delaying-the-load-event flag is set</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
async_test(function(t)
{
    var a = document.createElement("audio");
    a.foo = "bar";
    a.appendChild(document.createElement("source"));
    a.onloadstart = t.step_func(function(e)
    {
        assert_equals(e.target.foo, "bar");
        t.done();
    });
    a = null;
    gc();
});
</script>
